fix: panics in oauth providers
We were just checking each item in the slice of interfaces if they were
nil. The interface points to the data of the encapsulated type and an
itable. Checking if the item is nil, does not actually check if the
data encapsulated by the interface is nil, so the previous check wasn't
doing what we actually wanted. Instead we can modify the existing interface
to add a `Provided() bool` to check if the data is actually nil or not. This
should prevent the previous panic and issues with OAuth not working.