fix: Ingress objects showing progressing in ArgoCD
Similar to the previous change for Load Balancer services, in order to
make Ingresses not show as progressing in ArgoCD, we need to add a
status.loadBalancer.ingress object so that it appears healthy. This
adjusts the logic as well so that we don't do a n x m comparisons, but
rather n + m comparisions as this will be less for large values of n and
m. This also makes it so that we update the status for both ngrok
managed domains and non ngrok managed domains.

Ref: https://github.com/argoproj/gitops-engine/blob/72bcdda3f0a5b80432d9f72e5b30827a530ac349/pkg/health/health_ingress.go#L7-L16