add new mapping strategy (#612)
* implement collapse endpoints mapping strategy

Signed-off-by: Alice-Lilith <a.wasko@ngrok.com>

* change mapping strategy on existing translation tests

There is nothing wrong with these existing tests, but the default mapping-strategy is changing, so update the annotations on them
so that they test the endpoints-verbose strategy which is no longer the default.

Signed-off-by: Alice-Lilith <a.wasko@ngrok.com>

* update translation tests to expect set-vars instead of log

These translation tests needed to be updated to expect set-vars actions and expressions instead of log metadata.

Signed-off-by: Alice-Lilith <a.wasko@ngrok.com>

* update default mapping-strategy tests

These tests are using the default mapping strategy which has changed. They now expect
to not create any CloudEndpoint resources and collapse into one or more AgentEndpoint resources.

Signed-off-by: Alice-Lilith <a.wasko@ngrok.com>

* add additional collapse mapping strategy tests

Signed-off-by: Alice-Lilith <a.wasko@ngrok.com>

* implement both mapping strategies for LoadBalancer Services

Signed-off-by: Alice-Lilith <a.wasko@ngrok.com>

* fix parsing for mapping-strategy values

Signed-off-by: Alice-Lilith <a.wasko@ngrok.com>

---------

Signed-off-by: Alice-Lilith <a.wasko@ngrok.com>