fix(lb-services): Fix too frequent reconciles for lb-services (#694)
The change in #693 to fix a bug with Cloud/Agent Endpoint changes
not triggering re-reconciles had the adverse affect of watching
for all changes on cloud/agent endpoints and re-conciling too
frequently. We now use a better predicate filter on the service
to determine if it needs to be updated.

The status update was also not deterministic since the cloud and
agent endpoint status are different. We now correctly determine
the right object to use for status update.

In addition, we now also guard against trying to update the status
if it is already up to date.

Signed-off-by: Jonathan Stacks <jonstacks@users.noreply.github.com>